restructure for separate load & serve
